feat: Add tests, update notes implementation

+++ lib/features/document_scan/cubit/document_scanner_cubit.dart
@@ -54,7 +54,9 @@ class DocumentScannerCubit extends Cubit<DocumentScannerState> {
 
   Future<void> removeScan(File file) async {
     try {
-      await file.delete();
+      if (await file.exists()) {
+        await file.delete();
+      }
     } catch (error, stackTrace) {
       throw InfoMessageException(
         code: ErrorCode.scanRemoveFailed,
